Nested replies working
In a rather rudimentary fashion but with a bit of spit and polish they
ought to look nice. Unfortunately we manage the display state along
side the structure of the comments, may need to apply core.async to
make this less nasty. Or possibly just ignore it on the server
side (when that gets implemented).